+++ backend/src/controllers/auth.controller.ts
@@ -10,7 +10,7 @@ import User from '../interfaces/User';
 import { HttpStatusCode } from 'axios';
 import jwt from 'jsonwebtoken';
 import logger from '../utils/logger';
-import { errorMessages } from '../constants/constants';
+import { errorMessages } from '../utils/constants';
 
 export const register = async (req: Request, res: Response) => {
   try {
